BUGFIX: Standalone Core - move `AuthProviderFactoryInterface` to correct packgage
Neos\ContentRepository\Core MUST NOT have a dependency to Neos\ContentRepositoryRegistry

this was a hickup because at one point the cr factory did not need an AuthProviderFactoryInterface but the AuthProvider directly.

see b4e0c7829b61bc962986a898fcc10d929d0632a6
