+++ Neos.ContentRepository.Core/Classes/Factory/AuthProviderFactoryInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Neos\ContentRepositoryRegistry\Factory\AuthProvider;
+namespace Neos\ContentRepository\Core\Factory;
 
 use Neos\ContentRepository\Core\Feature\Security\AuthProviderInterface;
 use Neos\ContentRepository\Core\Projection\ContentGraph\ContentGraphReadModelInterface;
