+++ Neos.ContentRepository.TestSuite/Classes/Fakes/FakeAuthProviderFactory.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Neos\ContentRepository\TestSuite\Fakes;
 
+use Neos\ContentRepository\Core\Factory\AuthProviderFactoryInterface;
 use Neos\ContentRepository\Core\Feature\Security\AuthProviderInterface;
 use Neos\ContentRepository\Core\Projection\ContentGraph\ContentGraphReadModelInterface;
 use Neos\ContentRepository\Core\SharedModel\ContentRepository\ContentRepositoryId;
-use Neos\ContentRepositoryRegistry\Factory\AuthProvider\AuthProviderFactoryInterface;
 
 final class FakeAuthProviderFactory implements AuthProviderFactoryInterface
 {
